third_party/WebKit/LayoutTests/external/wpt/css/css-masking/parsing/mask-border-width-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-border-width with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-border-width">
<meta name="assert" content="mask-border-width supports the full grammar '[ <length-percentage> | <number> | auto ]{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("mask-border-width", "-10px");
test_valid_value("mask-border-width", "-10px -20");
test_valid_value("mask-border-width", "-10px -20 -30%");
test_valid_value("mask-border-width", "-10px -20 -30% auto");
</script>
</body>
</html>
